admin-webapp/requirements.txt
flask==2.3.3
werkzeug==2.3.7
google-cloud-storage==2.10.0
gunicorn==21.2.0